+++ execute/utils/redirection_parser/1_parse_redirections.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   1_parse_redirections.c                             :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: emgenc <emgenc@student.42istanbul.com.t    +#+  +:+       +#+        */
+/*   By: kuzyilma <kuzyilma@student.42istanbul.c    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/07/30 12:23:57 by emgenc            #+#    #+#             */
-/*   Updated: 2025/07/30 17:28:34 by emgenc           ###   ########.fr       */
+/*   Updated: 2025/07/31 19:02:29 by kuzyilma         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -108,4 +108,33 @@ char	**parse_redirs(t_split split, t_redir_fds *fds, t_shell *shell)
 	}
 	free_args(args);
 	return (clean_args);
+}
+
+int	process_single_redir(int i, t_redir_fds *fds,
+	t_shell *shell, t_heredoc_params *params)
+{
+	t_redirect_info	redirect_info;
+	int				rc;
+
+	redirect_info = get_redirect_info(params->args, i, shell);
+	if (redirect_info.redirect_type == 1 && fds->preprocessed_heredoc)
+	{
+		if (redirect_info.processed_filename != params->args[i + 1])
+			free(redirect_info.processed_filename);
+		return (0);
+	}
+	if (redirect_info.redirect_type == -1)
+	{
+		write(STDERR_FILENO, "syntax error: missing filename\n", 32);
+		return (-1);
+	}
+	if (!redirect_info.processed_filename)
+		return (-1);
+	params->delimiter = params->args[i + 1];
+	rc = handle_redirection_type(&redirect_info, fds, shell, params);
+	if (rc == -1)
+		return (after_handle(&redirect_info, params, i));
+	if (redirect_info.processed_filename != params->args[i + 1])
+		free(redirect_info.processed_filename);
+	return (0);
 }
