+++ main/minishell.h
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   minishell.h                                        :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: emgenc <emgenc@student.42istanbul.com.t    +#+  +:+       +#+        */
+/*   By: kuzyilma <kuzyilma@student.42istanbul.c    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/07/25 11:11:41 by emgenc            #+#    #+#             */
-/*   Updated: 2025/07/31 16:03:38 by emgenc           ###   ########.fr       */
+/*   Updated: 2025/07/31 20:08:20 by kuzyilma         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -202,15 +202,16 @@ void					remove_outer_par(char *start, char *end);
 int						delimiter_was_quoted(char *filename);
 void					clean_empties(t_split *split);
 int						validate_exec(char **args, char *executable);
-void					handle_error_and_exit(char **args, char *executable,
-							t_shell *shell);
 int						countchr_str(char *str, char c);
 int						countchr_not_quote(char *str, char c);
 void					sep_opt_arg(t_shell *shell);
 int						count_str_split(t_split split, const char *str,
 							int flag);
 char					*get_cut_indexs(t_split split);
+t_redirect_info			get_redirect_info(char **args, int i, t_shell *shell);
 void					cut_out_par(t_split *split);
+int						after_handle(t_redirect_info *redirect_info,
+							t_heredoc_params *params, int i);
 int						check_op_errs(t_split split);
 int						check_par_errs(t_split split);
 
@@ -223,8 +224,6 @@ int						set_env_var(char *var_name, char *value,
 int						replacevar(t_expand *holder, int flags);
 char					*expandvar_q(char *str, t_shell *shell);
 int						expand_dollar(t_expand *holder, t_shell *shell);
-void					init_expand_holder(t_expand *holder,
-							char *tilde_expanded);
 int						append_double_quotes(char *string);
 int						execute_args_array(t_split split, t_shell *shell);
 int						validate_redirection_syntax(t_split split);
@@ -234,7 +233,6 @@ char					**process_argument_expansion(char **args,
 int						execute_multi_command_pipeline(t_split *commands,
 							int cmd_count, t_shell *shell);
 t_redir_fds				get_fds(int *fd_values);
-int						has_parentheses_in_split(t_split split);
 
 int						is_builtin(char *cmd);
 int						execute_builtin(char **args, t_shell *shell);
@@ -276,14 +274,7 @@ int						execute_pipe_redir(t_split split, t_shell *shell);
 int						execute_single_command(char **args, t_shell *shell);
 int						count_pipes(t_split split);
 t_split					*split_by_pipes(t_split split, int *cmd_count);
-void					expand_command_args(char **args, t_shell *shell);
-int						handle_builtin_command(char **args, t_pipe_context *ctx,
-							t_shell *shell);
-int						handle_external_command(char **args,
-							t_pipe_context *ctx, t_shell *shell);
 void					execute_pipe_child(t_child_params *params);
-char					**execute_expanded_args_split(char *reconstructed,
-							char **args, t_shell *shell);
 void					execute_pipe_external_command(char **args,
 							t_shell *shell, t_pipe_cleaner *cleanup);
 void					setup_pipe_signals(void);
@@ -291,9 +282,7 @@ void					free_child_pipeline_memory(char **args, t_shell *shell,
 							t_pipe_cleaner *cleanup);
 void					execute_child_redir(t_child_redir_params *params);
 char					*remove_quotes_for_redirection(char *str);
-int						create_pipes_array(int ***pipes, int cmd_count);
 void					pipe_clean(int **pipes, int cmd_count);
-void					setup_child_redir(t_pipe_context *ctx);
 int						setup_pipe(t_split **commands,
 							int ***pipes, pid_t **pids, int cmd_count);
 void					clean_pipe(t_split *commands, int **pipes, pid_t *pids,
@@ -303,16 +292,10 @@ int						handle_question_mark(t_expand *holder, t_shell *shell);
 int						handle_special_chars(t_expand *holder);
 int						is_special_directory(char *cmd);
 void					write_path_error_message(char *cmd, char *message);
-void					update_command_filename(t_pipe_ctx *pipeline_ctx,
-							int i, int j, char *temp_filename);
-void					free_everything_in_child(char *delim,
-							t_heredoc_sub heresub, t_pipe_ctx *pipeline_ctx,
-							char *temp_filename);
 void					process_and_write_line(char *line, int pipe_fd,
 							t_shell *shell, int should_expand);
 void					exec_hdoc_chld(char *delimiter, int pipe_fd,
 							t_shell *shell, int should_expand);
-void					free_heredoc(t_shell *shell, t_heredoc_params *params);
 int						handle_escaped_quote(char *str, char *result,
 							int i, int j);
 int						should_copy_character(char c, int *states);
@@ -321,8 +304,6 @@ int						handle_redirection_type(t_redirect_info *info,
 							t_redir_fds *fds, t_shell *shell,
 							t_heredoc_params *params);
 char					*handle_quote_char(char *result, char quote_char);
-char					*handle_double_quote_case(char *result, int *i,
-							int *in_double_quotes, int in_single_quotes);
 char					*handle_segment_processing(char *result, char *str,
 							int *vars, t_shell *shell);
 int						clean_heredoc(char *processed_filename,
@@ -347,8 +328,6 @@ int						execute_with_redirections(t_split split,
 int						process_single_redir(int i, t_redir_fds *fds,
 							t_shell *shell, t_heredoc_params *params);
 char					*process_heredoc_delimiter(char *filename);
-char					*handle_double_quote_case(char *result, int *i,
-							int *in_double_quotes, int in_single_quotes);
 
 void					process_args_quotes(char **args, t_shell *shell);
 char					*expandvar_quoted(char *str, t_shell *shell);
@@ -364,7 +343,6 @@ int						check_single_par(t_split split);
 int						not_dir(char *path);
 void					write_notdir(char *filename);
 int						execute_builtin_with_redirect(t_redir_exec_ctx *ctx);
-void					write_line_to_pipe(int pipe_fd, char *line);
 t_pipe_cleaner			get_cleanup(t_split *commands, int **pipes,
 							pid_t *pids, int cmd_count);
 t_pipe_setup_ctx		get_pipe_ctx(int cmd_index, int cmd_count,
@@ -374,5 +352,10 @@ void					redirect_fds(int input_fd, int output_fd,
 void					close_all_pipes(int **pipes, int cmd_count);
 int						execute_external_with_redirect(t_redir_exec_ctx *ctx);
 void					setup_pipe_fds(t_pipe_setup_ctx *ctx);
+void					assign_params(t_heredoc_params *p, char **tokens,
+							char *delim);
+int						fd_closer(int *fd);
+int						exec_pipe_child(t_pipe_ctx *pipeline_ctx);
+int						preprocess_heredocs(t_pipe_ctx *ctx);
 
 #endif
