--- val.supp
@@ -1,48 +1,0 @@
-{
-   <Readline>
-   Memcheck:Leak
-   match-leak-kinds: reachable
-   ...
-   fun:readline
-   ...
-}
-{
-   <Readline/AddHistory>
-   Memcheck:Leak
-   match-leak-kinds: reachable
-   ...
-   fun:add_history
-   ...
-}
-{
-   <Bash>
-   Memcheck:Leak
-   match-leak-kinds: reachable
-   ...
-   fun:copy_command
-   ...
-}
-{
-   <Bash>
-   Memcheck:Leak
-   match-leak-kinds: reachable
-   ...
-   fun:initialize_job_control
-   ...
-}
-{
-   <Bash>
-   Memcheck:Leak
-   match-leak-kinds: reachable
-   ...
-   fun:execute_command_internal
-   ...
-}
-{
-   <Bash>
-   Memcheck:Leak
-   match-leak-kinds: reachable
-   ...
-   fun:initialize_shell_variables
-   ...
-}
